Interface created for user inside express
