+++ .gitignore
@@ -7,7 +7,10 @@
 /bazel-out
 
 # Node
+/server
 /server/node_modules
+/typescript/node_modules
+/typescript/dist
 /LoginForm/.angular
 /LoginForm/.vscode
 npm-debug.log
