+++ typescript/package.json
@@ -1,24 +1,23 @@
 {
-  "name": "server",
+  "name": "typescript",
   "version": "1.0.0",
   "description": "",
-  "main": "app.ts",
-  "type": "module",
+  "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.ts"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.12.6",
     "cors": "^2.8.5",
     "express": "^4.19.2",
     "nodemon": "^3.1.0",
     "typescript": "^5.4.4"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21"
+    "@types/cors": "^2.8.17"
   }
 }
